Add bar by default
